switched the order of the mHM calls in the objective function
in every call of mHM, L11_Qmod and other routing variables are freshly
allocated, if there is routing or not. So, if we call mHM for routing
before calling mHM for the other optimization values then L11_Qmod will
be all 0 in the restart files.
Switching the calls solves the problem. It should be tested if the
corresponding restart would work, or a parallel restart setup is needed.