adapted old tws objective function and fixed a bug in mhm_eval
in mhm_eval the timestep input of tws is needed and accidentally that
one of et was used. Since L1_data has to be loaded anyway,
writeOutCounter is now part of the optidata type. Maybe a better place
would be a new data type for the simulated data, or all goes together in
one type and is passed to mHM as such.