deleted redundant code
the variable name now also is part of the datastructure and set in the
mHM read config. With this change, and since the nTimeSteps can be
handled individually for each domain and isn't part of the reading
anymore, the reading of optional gridded data works for all
et/tws/sm/neutrons in the same way.