smhorizon_bugfix
This is a bug fix. In the old version we mistakenly correct tmp_rootfraccoef directly if it is not between 0 and 1, but we should actually make sure that FCnorm is always between 0 and 1. This will in turn also secure that tmp_rootfraccoef is between 0 and 1.
Case 10 and 12 files are also updated accordingly. All test cases in CYGWIN-Python are successful.